Refactor plugins
Problem: .eleventy.js is full of configuration, so much so that it's hard to navigate.

Precedent: The Handbook groups much of its config into a HandbookPlugin https://github.com/18F/handbook/blob/main/.eleventy.js#L10-L15

This commit refactors all of the configuration into a clearer organizational scheme, importing collections of functionality as a single 18F plugin (named EighteenF because you can't start variable names with numbers).
