+++ config/collections/services.js
@@ -1,0 +1,1 @@
+module.exports = async (collection) => collection.getFilteredByGlob('content/pages/projects/services/*.md')
