+++ config/data_extensions/index.js
@@ -1,0 +1,10 @@
+const { parse } = require('csv-parse/sync'); /* eslint-disable-line import/no-unresolved */
+const yaml = require('js-yaml')
+
+// Allow YAML and CSV data to be used in the _data dir
+const dataExtensionsPlugin = (eleventyConfig) => {
+  eleventyConfig.addDataExtension('csv', (contents) => parse(contents, { columns: true, skip_empty_lines: true }));
+  eleventyConfig.addDataExtension('yml, yaml', contents => yaml.load(contents))
+}
+
+module.exports = dataExtensionsPlugin;
