+++ config/filters/asAbsoluteUrl.js
@@ -1,0 +1,7 @@
+const fs = require('fs');
+const yaml = require('js-yaml');
+
+module.exports = async (relativeUrl) => {
+  const { host } = yaml.load(fs.readFileSync('./_data/site.yaml', 'utf8'))
+  return new URL(relativeUrl, host).href
+}
