+++ config/filters/capitalize.js
@@ -1,0 +1,1 @@
+module.exports = async (value) => value.charAt(0).toUpperCase() + value.slice(1)
