+++ config/filters/dateToXMLSchema.js
@@ -1,0 +1,3 @@
+const { dateObject } = require('../../lib/date')
+
+module.exports = async (date) => (await dateObject(date)).toISOString();
