+++ config/filters/filterTagList.js
@@ -1,0 +1,3 @@
+module.exports = async (tags) => (tags || []).filter(
+  (tag) => ['all', 'nav', 'post', 'posts'].indexOf(tag) === -1,
+)
