+++ config/filters/limit.js
@@ -1,0 +1,1 @@
+module.exports = async (arr, limitAmount) => arr.slice(0, limitAmount)
