+++ config/filters/markdownify.js
@@ -1,0 +1,6 @@
+const MarkdownIt = require('markdown-it');
+
+module.exports = async (content) => {
+  const md = new MarkdownIt({ html: true });
+  return md.render(content);
+};
