+++ config/filters/matching.js
@@ -1,0 +1,1 @@
+module.exports = async (collection, author) => collection.filter((post) => post.data.authors.includes(author))
