+++ config/filters/min.js
@@ -1,0 +1,2 @@
+// Return the smallest number argument
+module.exports = async (...numbers) => Math.min.apply(null, numbers);
