+++ config/filters/readableDate.js
@@ -1,0 +1,3 @@
+const { DateTime } = require('luxon');
+
+module.exports = async (dateObj) => DateTime.fromJSDate(dateObj, { zone: 'utc' }).toFormat
