+++ config/filters/teamLink.js
@@ -1,0 +1,6 @@
+const { findPerson, fullName } = require('../../lib/team');
+
+module.exports = async (slug) => {
+  const name = fullName(findPerson(slug))
+  return `<a href="/author/${slug.toLowerCase()}/">${name}</a>`
+}
