+++ config/shortcodes/image.js
@@ -1,0 +1,3 @@
+const imageWithClass = require('./imageWithClass')
+
+module.exports = async (src, alt) => await imageWithClass(src, '', alt)
