+++ config/shortcodes/uswdsIcon.js
@@ -1,0 +1,4 @@
+module.exports = async (name) => `
+  <svg class="usa-icon" aria-hidden="true" role="img">
+    <use xlink:href="#svg-${name}"></use>
+  </svg>`
